requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
Django==4.2.9
django-cleanup==8.0.0
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
reportlab==4.0.8
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
